Various minor cleanups to fix some things introduced in recent
Estimated hours taken: 2

Various minor cleanups to fix some things introduced in recent
changes to the runtime.

mercury_context.c:
mercury_context.h:
mercury_regorder.h:
mercury_regs.h:
mercury_thread.c:
mercury_thread.h:
mercury_type_info.c:
mercury_engine.h:
mercury_engine.c:
	Fix up the layout in a few places.  For multi-line macros
	defined using `do { ... } while (0)', put the `do {' on a line
	of its own.
	Change some names to conform to our standard naming
	convention and to avoid name clashes:
		s/Context/MR_Context/g
		s/CONTEXT/MR_context_struct/g
		s/SYNCTERM/MR_sync_term/g
		s/sp/MR_sp/g
		s/curfr/MR_curfr/g
		s/maxfr/MR_maxfr/g

mercury_engine.h:
mercury_engine.c:
	Fix a couple of bugs that caused it to not compile in certain
	grades.
	Delete some obsolete comments.  Add some new ones and rearrange
	some of the code to make it more readable.

mercury_type_info.c:
	Cast values used in switch statements to (int), to avoid
	a warning from gcc.

+++ runtime/mercury_type_info.c
@@ -545,11 +545,11 @@ MR_make_type_info(Word *term_type_info, Word *arg_pseudo_type_info,
 enum MR_DataRepresentation
 MR_categorize_data(Word functors_indicator, Word layout_entry)
 {
-	switch (functors_indicator) { 
+	switch ((int) functors_indicator) { 
 		case MR_TYPEFUNCTORS_ENUM: 
 			return MR_DATAREP_ENUM;
 		case MR_TYPEFUNCTORS_DU: 
-			switch (tag(layout_entry)) {
+			switch ((int) tag(layout_entry)) {
 				case TYPELAYOUT_SIMPLE_TAG:
 					return MR_DATAREP_SIMPLE;
 				case TYPELAYOUT_COMPLICATED_TAG:
